fix: 增加菜单hidden single功能
